# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PP App API Documentation",
        "description": "This is the API documentation for the PP application.",
        "contact": {
            "email": "support@pp-app.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "PP App API Server"
        }
    ],
    "paths": {
        "/api/appointments": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get list of appointments",
                "description": "Returns list of appointments",
                "operationId": "add744e6b5d26e8ac07c106df05ddd95",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Appointment"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Create a new appointment",
                "description": "Create a new appointment",
                "operationId": "098ba7006404045ebc44117b157d17a2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Appointment"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Appointment created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Appointment"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/appointments/{id}": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get appointment details",
                "description": "Returns a single appointment",
                "operationId": "d4495d90592322051b3432ec3c7008f3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Appointment"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Update an appointment",
                "description": "Update an appointment",
                "operationId": "7a66357d0024f1bb9c0cf665282eb3a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Appointment"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Appointment updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Appointment"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Delete an appointment",
                "description": "Deletes an appointment",
                "operationId": "319ced5e50b9be431b2ef6b9703acf84",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                }
            }
        },
        "/api/health-records": {
            "get": {
                "tags": [
                    "Health Records"
                ],
                "summary": "List all health records",
                "description": "Retrieve a list of all health records.",
                "operationId": "411ff914e6795f0f8f3cfd7d66f9a66e",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HealthRecord"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Health Records"
                ],
                "summary": "Create a new health record",
                "description": "Store a new health record in the database.",
                "operationId": "636fcd3a20ddfb2d46e8db1e5fbc33ed",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HealthRecord"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Record created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HealthRecord"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/health-records/{id}": {
            "get": {
                "tags": [
                    "Health Records"
                ],
                "summary": "Get a specific health record",
                "description": "Retrieve a specific health record by ID.",
                "operationId": "d52cc7c77dc5f52ab16eeb8735c3abfd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HealthRecord"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Health Records"
                ],
                "summary": "Update an existing health record",
                "description": "Update a specific health record by ID.",
                "operationId": "177f03c79fc3431b38736962b5be5685",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HealthRecord"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Record updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HealthRecord"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Health Records"
                ],
                "summary": "Delete a health record",
                "description": "Remove a specific health record by ID from the database.",
                "operationId": "4f18c7bb85451c0b98388d2c3855fd88",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Record deleted successfully"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Appointment": {
                "title": "Appointment",
                "description": "Appointment model",
                "required": [
                    "patient_id",
                    "healthcare_professional_id",
                    "appointment_date",
                    "status"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the appointment",
                        "type": "integer",
                        "example": 1
                    },
                    "patient_id": {
                        "description": "ID of the patient",
                        "type": "integer",
                        "example": 42
                    },
                    "healthcare_professional_id": {
                        "description": "ID of the healthcare professional",
                        "type": "integer",
                        "example": 7
                    },
                    "appointment_date": {
                        "description": "Date and time of the appointment",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-24T14:30:00Z"
                    },
                    "status": {
                        "description": "Status of the appointment",
                        "type": "string",
                        "example": "confirmed"
                    },
                    "created_at": {
                        "description": "Timestamp when the appointment was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-23T14:30:00Z"
                    },
                    "updated_at": {
                        "description": "Timestamp when the appointment was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-23T15:30:00Z"
                    }
                },
                "type": "object"
            },
            "HealthRecord": {
                "title": "Health Record",
                "required": [
                    "patient_id",
                    "description",
                    "type"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "patient_id": {
                        "description": "ID of the patient",
                        "type": "integer",
                        "example": 42
                    },
                    "description": {
                        "type": "string",
                        "example": "Routine checkup"
                    },
                    "type": {
                        "type": "string",
                        "example": "diagnosis"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-20T18:25:43.511Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-21T18:25:43.511Z"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Appointments",
            "description": "API Endpoints for Managing Appointments"
        },
        {
            "name": "Health Records",
            "description": "API Endpoints for Health Records"
        }
    ]
}